change: 탭 목록 구조 변경
