+++ components/Tab/TabList/index.tsx
@@ -1,47 +1,11 @@
-import { cn } from "@/util/cn";
-import { useContext } from "react";
-import { TabContext } from "../context";
-
-type TabItem = {
-  label: string;
-  id: "T" | "I";
-};
-
 type Props = {
-  tabs: Array<TabItem>;
-  defaultValue?: string;
+  children: React.ReactNode | React.ReactNode[];
 };
 
-export default function TabList({ tabs }: Props) {
-  const context = useContext(TabContext);
-
-  if (!context) {
-    return null;
-  }
-
-  const tabClick = (id: string) => {
-    context.setSelectTab(id);
-  };
-
+export default function TabList({ children }: Props) {
   return (
     <div className="w-full bg-white/90">
-      <div className="flex items-center gap-1 p-1">
-        {tabs.map((tab) => (
-          <button
-            key={tab.id}
-            role="tab"
-            className={cn(
-              "min-w-[140px] p-1 text-sm text-gray-400 transition",
-              tab.id === context.selectTab &&
-                "rounded bg-white font-bold text-gray-800",
-            )}
-            value={tab.id}
-            onClick={() => tabClick(tab.id)}
-          >
-            {tab.label}
-          </button>
-        ))}
-      </div>
+      <div className="flex items-center gap-1 p-1">{children}</div>
     </div>
   );
 }
