+++ components/Tab/index.tsx
@@ -1,3 +1,4 @@
 export { default as TabContainer } from "./TabContainer";
 export { default as TabList } from "./TabList";
+export { default as TabItem } from "./TabItem";
 export { default as TabContents } from "./TabContents";
